Crate wio

A zero overhead Windows I/O library

Structs

AcceptAddrs
The parsed return value of AcceptAddrsBuf.

AcceptAddrsBuf
A type to represent a buffer in which an accepted socket's address will be stored.

CompletionPort
A handle to an Windows I/O Completion Port.

CompletionStatus
A status message received from an I/O completion port.

SocketAddrBuf
A type to represent a buffer in which a socket address will be stored.

Traits

TcpBuilderExt
Additional methods for the TcpBuilder type in the net2 library.

TcpListenerExt
Additional methods for the TcpListener type in the standard library.

TcpStreamExt
Additional methods for the TcpStream type in the standard library.

UdpSocketExt
Additional methods for the UdpSocket type in the standard library.