Struct wio::iocp::CompletionStatus

pub struct CompletionStatus(_);

A status message received from an I/O completion port.

These statuses can be created via the new or empty constructors and then provided to a completion port, or they are read out of a completion port. The fields of each status are read through its accessor methods.

Methods

impl CompletionStatus

fn new(bytes: u32, token: usize, overlapped: *mut OVERLAPPED) -> CompletionStatus

Creates a new completion status with the provided parameters.

This function is useful when creating a status to send to a port with the post method. The parameters are opaquely passed through and not interpreted by the system at all.

fn zero() -> CompletionStatus

Creates a new "zero" completion status.

This function is useful when creating a stack buffer or vector of completion statuses to be passed to the get_many function.

fn bytes_transferred(&self) -> u32

Returns the number of bytes that were transferred for the I/O operation associated with this completion status.

fn token(&self) -> usize

Returns the completion key value associated with the file handle whose I/O operation has completed.

A completion key is a per-handle key that is specified when it is added to an I/O completion port via add_handle or add_socket.

fn overlapped(&self) -> *mut OVERLAPPED

Returns a pointer to the OVERLAPPED structure that was specified when the I/O operation was started.