Module wio::pipe

Named pipes

Structs

AnonRead
Readable half of an anonymous pipe.

AnonWrite
Writable half of an anonymous pipe.

NamedPipe
A named pipe that can accept connections.

NamedPipeBuilder
A builder structure for creating a new named pipe.

Functions

anonymous
Creates a new anonymous in-memory pipe, returning the read/write ends of the pipe.

connect
A convenience function to connect to a named pipe.