Function wio::pipe::connect

pub fn connect<A: AsRef<OsStr>>(addr: A) -> Result<File>

A convenience function to connect to a named pipe.

This function will block the calling process until it can connect to the pipe server specified by addr. This will use NamedPipe::wait internally to block until it can connect.